+++ pipeline-multithreaded-trade-processing/trading-app/src/main/java/io/reactivestax/service/TradeProcessorService.java
@@ -7,7 +7,7 @@ import java.io.IOException;
 public interface TradeProcessorService {
     void processTrade(String tradeId) throws InterruptedException, IOException;
 
-    JournalEntry journalEntryTransaction(String[] payloadArr, int tradeId);
+    JournalEntry journalEntryTransaction(String[] payloadArr, Long tradeId);
 
     void positionTransaction(JournalEntry journalEntry);
 }
